working per contractservice view

+++ src/lib/server/db/schema/contract.ts
@@ -1,4 +1,4 @@
-import { mysqlTable, varchar, double, serial, uniqueIndex, int } from 'drizzle-orm/mysql-core';
+import { mysqlTable, varchar, serial, uniqueIndex, int } from 'drizzle-orm/mysql-core';
 import { relations } from 'drizzle-orm';
 import { contractService, employee, contractType, salaryType } from './';
 
@@ -7,7 +7,7 @@ export const contract = mysqlTable(
 	{
 		id: serial('id').primaryKey(),
 		number: varchar('number', { length: 191 }).notNull(),
-		fixedSalary: double('fixedSalary'),
+		fixedSalary: int('fixedSalary'),
 		employeeId: int('employeeId').notNull(),
 		contractTypeId: int('contractTypeId').notNull(),
 		salaryTypeId: int('salaryTypeId')
